chore: update typing
