+++ src/frontend/screens/ObservationCreate/ObservationCreateSaveButton.tsx
@@ -28,7 +28,7 @@ import {
 import {
   isUnsavedAudioAttachment,
   isUnsavedPhotoAttachment,
-} from '../../lib/attachmentTypeCheckDraftStore';
+} from '../../lib/attachmentTypeChecks';
 
 const MAXIMUM_ACCURACY = 10;
