+++ src/frontend/screens/ObservationEdit/index.tsx
@@ -36,7 +36,7 @@ import {
 import {
   isUnsavedAudioAttachment,
   isUnsavedPhotoAttachment,
-} from '../../lib/attachmentTypeCheckDraftStore';
+} from '../../lib/attachmentTypeChecks';
 
 const m = defineMessages({
   observation: {
